reddit account linking and unlinking implemented

+++ site/app/views.py
@@ -1,12 +1,13 @@
-from app import app, db, login_manager, bcrypt
+from app import app, db, login_manager, bcrypt, reddit
 from app.models import User, Imgur_User, Reddit_User
 from flask import render_template, request, redirect, url_for, session, flash
 from flask.ext.login import LoginManager, login_required, login_user, logout_user, current_user
 import datetime
 
 import app.imgur_backend.imgur_controller as im_control
+import app.reddit_backend.reddit_helper as r_h
 
-
+# TODO: Unique instance of Reddit Object per user
 
 login_manager.login_view = "login"
 
@@ -24,17 +25,16 @@ def login():
     For POSTs, login user by processing form
     '''
     # If user is logged in, redirect to home immediately
-    if(current_user):
+    if(current_user.is_active):
         return redirect(url_for('home'))
     # Otherwise, handle GET and POST normally
     if (request.method == 'GET'):
         return render_template('login.html', page='login')
     elif (request.method == 'POST'):
-        print(request.form)
         user = User.query.filter_by(username=request.form['username']).first()
         if(user is None):
             flash("Username and password combination not found", 'error')
-            return render_template('login.html', page='login')
+            return redirect(url_for("login"))
         elif(bcrypt.check_password_hash(user.pwd, request.form['password'])):
             user.authenticated = True
             db.session.add(user)
@@ -43,6 +43,9 @@ def login():
             if('remember' in request.form):
                 remember_login=True
             login_user(user, remember=remember_login)
+            # Re-establish OAuth for reddit object if it exists
+            reddit_oauth = r_h.reestablish_oauth(reddit,
+                            current_user.reddit_user)
             return redirect(url_for('home'))
 
 @app.route('/register/process', methods=['POST'])
@@ -54,7 +57,7 @@ def register_process():
         if(request.form['confirm-password'] == request.form['password']):
             hashed = bcrypt.generate_password_hash(request.form['password'])
             user = User(username=request.form['username'], pwd=hashed, email=request.form['email'])
-            # Register user, log-in, and redirect
+            # Register, log-in, and redirect user
             user.authenticated = True
             db.session.add(user)
             db.session.commit()
@@ -68,16 +71,19 @@ def register_process():
 @login_required
 def logout():
     user = current_user
-    user.authenticated = False
-    db.session.add(user)
-    db.session.commit()
-    logout_user()
+    if(user):
+        user.authenticated = False
+        db.session.add(user)
+        db.session.commit()
+        logout_user()
     return redirect(url_for('login'))
 
 @app.route('/home')
 @login_required
 def home():
-    return render_template('home.html', page='home')
+    return render_template('home.html',
+            page='home',
+            in_title = "Home")
 
 @app.route('/about')
 def about():
@@ -87,27 +93,40 @@ def about():
 def contact():
     return 'Contact Page'
 
+@app.route('/recover')
+def recover_account():
+    # TODO: Account recovery
+    return 'Account Recovery Page'
+
 @app.route('/account')
 @login_required
 def account():
     # TODO: Delete user
+    # TODO: Account page
     return "%s's profile" % current_user.username
 
 @app.route('/account/imgur')
 @login_required
-def imgur_account():
-    return render_template("imgur_account.html", request_url=im_control.get_request_pin_url())
+def account_imgur():
+    return render_template("account_imgur.html",
+            page="account_imgur",
+            in_title="Account | Imgur",
+            request_url=im_control.get_request_pin_url())
     
 @app.route('/account/imgur/link', methods=['POST'])
 @login_required
-def imgur_account_link():
+def account_imgur_link():
     if ('user_pin' in request.form):
-        # Exchange for tokens
+        # Check if current_user has existing Imgur account
+        im_usr = Imgur_User.query.filter_by(user_id=current_user.id).first()
+        if(im_usr):
+            # There is existing account, stop and return
+            flash("There is an Imgur account already linked to your current account", "danger")
+            return redirect(url_for("account_imgur"))
+        # There is no account linked, continue to try to link
         response = im_control.exchange_pin_for_tokens(request.form['user_pin'])
         if('success' in response):
             if (response['success'] == True):
-                # TODO: Unlink possible remaining imgur_user
-
                 # Create new imgur_user
                 im_usr = Imgur_User(username=response['username'],
                         access_token=response['access_token'],
@@ -118,47 +137,90 @@ def imgur_account_link():
                 flash("Imgur Account Linked", 'success')
             elif (response['success'] == False):
                 flash(response['error'], 'danger')
-            return redirect(url_for("imgur_account"))
+            return redirect(url_for("account_imgur"))
         else: 
             # Internal error in imgur_backend handling
             return 'Unhandled server error'
-        # TODO: Create Imgur_User and bind to current_user account
     else:
         flash("Bad Input", 'danger')
-        return redirect(url_for("imgur_account"))
+        return redirect(url_for("account_imgur"))
 
 @app.route('/account/imgur/unlink', methods=['POST'])
 @login_required
-def imgur_account_unlink():
-    # TODO: Unbind imgur_user to User
+def account_imgur_unlink():
+    # Remove Imgur_User from database
     if ('submit' in request.form):
         im_usr = current_user.imgur_user
-        db.session.delete(im_usr)
-        db.session.commit()
-        flash("Imgur Account Unlinked", "success")
-    return redirect(url_for("imgur_account"))
+        if(im_usr):
+            db.session.delete(im_usr)
+            db.session.commit()
+            flash("Imgur Account Unlinked", "warning")
+        else:
+            flash("Imgur account not found", "danger")
+    return redirect(url_for("account_imgur"))
 
-@app.route('/recover')
-def recover_account():
-    # TODO: Account recovery
-    return 'Account Recovery Page'
+@app.route('/account/reddit')
+@login_required
+def account_reddit():
+    return render_template('account_reddit.html',
+            page='account_reddit',
+            in_title="Account | Reddit",
+            request_url = r_h.get_authorize_url(reddit))
+
+@app.route('/account/reddit/link', methods=['GET'])
+@login_required
+def account_reddit_link():
+    print(request.args)
+    if ('state' in request.args and 'code' in request.args):
+        cur_state = request.args.get('state')
+        # TODO: Check states
+        cur_code = request.args.get('code')
+        # TODO: Fail get_access_information gracefully
+        # TODO: Unlink before relink
+        info = reddit.get_access_information(cur_code)
+        if('access_token' in info and 'refresh_token' in info):
+            usr = reddit.get_me()
+            reddit_usr = Reddit_User(username=usr.name,
+                        access_token=info['access_token'],
+                        refresh_token=info['refresh_token'],
+                        user_id=current_user.id)
+            db.session.add(reddit_usr)
+            db.session.commit()
+            flash("Reddit Account Linked", 'success')
+            return redirect(url_for('account_reddit'))
+    else:
+        return redirect(url_for('account_reddit'))
+
+@app.route('/account/reddit/unlink', methods=['POST'])
+@login_required
+def account_reddit_unlink():
+    # Delete Reddit_User from database
+    if ('submit' in request.form):
+        reddit_usr = current_user.reddit_user
+        if (reddit_usr):
+            reddit.clear_authentication()
+            db.session.delete(reddit_usr)
+            db.session.commit()
+            flash("Reddit Account Unlinked", "warning")
+        else:
+            flash("Reddit Account Not Found", "danger")
+    return redirect(url_for("account_reddit"))
 
 @app.route('/upload_and_post/process', methods=['POST'])
 @login_required
 def upload_and_post():
     if(request.method == 'POST'):
         # TODO: Sanitize inputs
-        # TODO: Reddit Posting Portion
         response = im_control.image_upload(current_user.imgur_user, request.form['img_url'])
         if('success' in response):
             if (response['success'] == True):
+                # Image Uploaded
                 flash(response['imgur_url'], 'success')
+                # TODO: Reddit Posting and Commenting Portion
+                # TODO: Captcha Handling
             elif (response['success'] == False):
                 flash(response['error'], 'danger')
         else: # Internal error in imgur_backend handling
             return 'Unhandled server error'
         return redirect(url_for("home"))
 
-
-if __name__ == "__main__":
-    app.run(debug=True)
